// types/Miner.ts
export type Miner = {
    id: string | null;
    miner_tag: string | null;
    email: string | null;
    phone: string | null;
    walletAddress: string | null;
    name: string;
    profilePicture: string| null;
    colorOne: string | null;
    colorTwo: string | null;
    colorThree: string | null;
    direction: string | null;
    birthday: string | null;
    bio: string | null;
    instagram: string | null;
    twitter: string | null;
    createdAt: string | null;
    updatedAt: string | null;
}